Modularize `content.ts`
The goal: No code executed in module scope. As few variables as
possible in module scope. This will have several major benefits.

First, initialization will much more tightly controlled. Init
currently happens as modules are loaded, which means we have
uncontrollable race conditions (due to `import`s being hoisted to the
opt of each module where we can't synchronize them) and all sorts of
fun debugging issues.

Second, it allows modules to be included without initialization code
being executed. This will mean you can include your module as many
times as you want and from wherever you want without, for example,
adding infinitely-recursing commandline iframes because your operation
in the commandline frame script wanted some module that eventually
pulls in the content script which creates another commandline iframe.

Third, it means we regain encapsulation. Right now we have a ton of
global variables being exported from various modules where other
modules can read and write them freely. This is *bad*. This means that
our modules are tightly coupled and global state can have far-reaching
consequences that are hard to debug and even harder to refactor. This
*needs* to change. The way to fix it is to tighten variable scopes and
carefully design hierarchies and accessors.

This is nowhere near complete, even for this one function. I cleaned
up a lot of the simple stuff in `content.ts` but there are two very
hard parts remaining: contentState and the excmd controller. Both of
these are global variables that get sent everywhere and they are both
going to be huge pains to fix. It's *simple*, just a ton of editing:
the state needs to live in the content core and it needs to be passed
down to the things that need it.
