+++ src/lib/controller.ts
@@ -3,28 +3,60 @@ import { parser as exmode_parser } from "@src/parsers/exmode"
 
 const logger = new Logger("controller")
 
-let stored_excmds: any
-export function setExCmds(excmds: any) {
-    stored_excmds = excmds
+
+export class ExcmdAccepter {
+    private last_ex_str: string = ""
+
+    constructor(
+        // This should really be something like this:
+        // 
+        //   readonly excmds: { [ns: string]: { [excmd: string]: (...any) => any } },
+        //
+        // But excmds.ts still has a ton of variables and stuff in it,
+        // so we can't do that and have to leave it as `any`.
+        readonly excmds: { [ns: string]: { [excmd: string]: any } },
+    ) { }
+
+    async acceptExCmd(exstr: string): Promise<any> {
+        // TODO: Errors should go to CommandLine.
+        try {
+            const [func, args] = exmode_parser(exstr, this.excmds)
+            // Stop the repeat excmd from recursing.
+            if (func !== this.excmds.repeat) this.last_ex_str = exstr
+            try {
+                return await func(...args)
+            } catch (e) {
+                // Errors from func are caught here (e.g. no next tab)
+                logger.error("controller in excmd: ", e)
+            }
+        } catch (e) {
+            // Errors from parser caught here
+            logger.error("controller while accepting: ", e)
+        }
+    }
+
+    repeat(n = 1, ...exstr: string[]) {
+        let cmd = this.last_ex_str
+        if (exstr.length > 0) cmd = exstr.join(" ")
+        logger.debug("repeating " + cmd + " " + n + " times")
+        for (let i = 0; i < n; i++) {
+            this.acceptExCmd(cmd)
+        }
+    }
 }
 
-export let last_ex_str = ""
-
-/** Parse and execute ExCmds */
-export async function acceptExCmd(exstr: string): Promise<any> {
-    // TODO: Errors should go to CommandLine.
-    try {
-        const [func, args] = exmode_parser(exstr, stored_excmds)
-        // Stop the repeat excmd from recursing.
-        if (func !== stored_excmds.repeat) last_ex_str = exstr
-        try {
-            return await func(...args)
-        } catch (e) {
-            // Errors from func are caught here (e.g. no next tab)
-            logger.error("controller in excmd: ", e)
-        }
-    } catch (e) {
-        // Errors from parser caught here
-        logger.error("controller while accepting: ", e)
-    }
+// TODO: Propagate the excmdAccepter far enough down that we can get
+// rid of these globals
+export let root_accepter: ExcmdAccepter
+export function setExCmds(excmds: any) {
+    root_accepter = new ExcmdAccepter(excmds)
+}
+export function setGlobalAccepter(new_accepter: ExcmdAccepter) {
+    root_accepter = new_accepter
+}
+export async function acceptExCmd(exstr: string) {
+    root_accepter.acceptExCmd(exstr)
+}
+export async function repeat(n, ...exstr: string[]) {
+    root_accepter.repeat(n, ...exstr)
 }
