+++ src/newtab.ts
@@ -2,6 +2,7 @@
 
 import * as Messaging from "@src/lib/messaging"
 import * as config from "@src/lib/config"
+import * as excmds_content from "@src/.excmds_content.generated"
 
 // These functions work with the elements created by tridactyl/scripts/newtab.md.sh
 function getChangelogDiv() {
@@ -48,5 +49,18 @@ window.addEventListener("load", _ => {
 window.addEventListener("load", _ => {
     if (config.get("update", "nag") === true) {
         Messaging.message("controller_background", "acceptExCmd", ["updatecheck auto_polite"])
+    }
+})
+
+config.getAsync("newtab").then(newtab => {
+    if (newtab !== "about:blank") {
+        if (newtab) {
+            excmds_content.open_quiet(newtab)
+        } else {
+            document.body.style.height = "100%"
+            document.body.style.opacity = "1"
+            document.body.style.overflow = "auto"
+            document.title = "Tridactyl Top Tips & New Tab Page"
+        }
     }
 })
